+++ app/Models/NewsFederalRiding.php
@@ -13,14 +13,21 @@ class NewsFederalRiding extends Model {
       'news_province_id',
       'population',
       'area',
-      'representation',
+      'representative',
       'geo_coordinates',
+      'geo_shape_json',
       'historical_data',
       'economic_indicators',
-      'date_founded'
+      'date_founded',
+      'year_updated',
+      'ed_code',
   ];
 
   public function province() {
     return $this->belongsTo(NewsProvince::class);
+  }
+
+  public function cities() {
+    return $this->hasMany(NewsCity::class);
   }
 }