+++ config/app.php
@@ -198,6 +198,7 @@ return [
         App\Providers\FortifyServiceProvider::class,
         App\Providers\JetstreamServiceProvider::class,
         Jenssegers\Agent\AgentServiceProvider::class,
+//        Maatwebsite\Excel\ExcelServiceProvider::class,
 
     ],