+++ database/migrations/2024_01_20_112639_rename_news_posts_to_news_stories.php
@@ -13,7 +13,25 @@ return new class extends Migration
      */
   public function up()
   {
+    Schema::table('news_posts', function (Blueprint $table) {
+      $table->dropUnique('news_posts_slug_unique');
+      $table->dropUnique('news_posts_title_unique');
+      $table->dropForeign(['user_id']);
+      $table->dropForeign(['image_id']);
+      $table->dropForeign(['status']);
+      $table->dropForeign(['video_id']);
+    });
+
     Schema::rename('news_posts', 'news_stories');
+
+    Schema::table('news_stories', function (Blueprint $table) {
+      $table->unique('slug', 'news_stories_slug_unique');
+      $table->unique('title', 'news_stories_title_unique');
+      $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+      $table->foreign('image_id')->references('id')->on('images');
+      $table->foreign('status')->references('id')->on('news_statuses')->onDelete('set null');
+      $table->foreign('video_id')->references('id')->on('videos')->onDelete('set null');
+    });
   }
 
     /**
@@ -23,6 +41,24 @@ return new class extends Migration
      */
   public function down()
   {
+    Schema::table('news_stories', function (Blueprint $table) {
+      $table->dropUnique('news_stories_slug_unique');
+      $table->dropUnique('news_stories_title_unique');
+      $table->dropForeign(['user_id']);
+      $table->dropForeign(['image_id']);
+      $table->dropForeign(['status']);
+      $table->dropForeign(['video_id']);
+    });
+
     Schema::rename('news_stories', 'news_posts');
+
+    Schema::table('news_posts', function (Blueprint $table) {
+      $table->unique('slug', 'news_posts_slug_unique');
+      $table->unique('title', 'news_posts_title_unique');
+      $table->foreign('user_id')->references('id')->on('users')->onDelete('set null');
+      $table->foreign('image_id')->references('id')->on('images');
+      $table->foreign('status')->references('id')->on('news_statuses')->onDelete('set null');
+      $table->foreign('video_id')->references('id')->on('videos')->onDelete('set null');
+    });
   }
 };