+++ database/migrations/2024_01_23_063433_remove_unique_constraint_from_movie_category_subs_name.php
@@ -26,7 +26,8 @@ return new class extends Migration
     public function down()
     {
       Schema::table('movie_category_subs', function (Blueprint $table) {
-        $table->unique('name');
+        // don't reverse this, if there is data in the table with non-unique names
+        // it will cause a problem. (tec21)
       });
     }
 };