+++ database/migrations/2024_01_23_092545_create_news_federal_ridings_table.php
@@ -16,15 +16,17 @@ return new class extends Migration
       Schema::create('news_federal_ridings', function (Blueprint $table) {
         $table->id();
         $table->string('name');
-        $table->unsignedBigInteger('news_province_id');
-        $table->foreign('news_province_id')->references('id')->on('news_provinces');
+        $table->foreignId('news_province_id')->references('id')->on('news_provinces');
         $table->integer('population')->nullable();
         $table->float('area')->nullable();
         $table->string('representative')->nullable();
         $table->string('geo_coordinates')->nullable();
+        $table->json('geo_shape_json')->nullable();
         $table->text('historical_data')->nullable();
         $table->string('economic_indicators')->nullable();
         $table->date('date_founded')->nullable();
+        $table->integer('year_updated')->nullable();
+        $table->integer('ed_code')->nullable();
         $table->timestamps();
       });
     }